requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
PyPDF2
python-docx
Pillow
pytesseract
transformers
sentence-transformers
chromadb
psycopg2-binary
sqlalchemy
python-dotenv
langchain
langchain-community
langchain-core
jieba